Finished applying checkstyle to main folder.
