+++ build.gradle
@@ -17,9 +17,10 @@ repositories {
 }
 
 dependencies {
-//    compile group: 'ch.qos.logback', name: 'logback-classic', version:'1.2.3'
-//    compile group: 'org.slf4j', name: 'slf4j-api', version:'1.7.25'
-    testCompile group: 'org.projectlombok', name: 'lombok', version:'1.16.16'
+    compile group: 'ch.qos.logback', name: 'logback-classic', version:'1.2.3'
+    compile group: 'org.slf4j', name: 'slf4j-api', version:'1.7.25'
+    compile group: 'org.projectlombok', name: 'lombok', version:'1.16.16'
+
     testCompile group: 'org.hamcrest', name: 'hamcrest-core', version:'1.3'
     testCompile group: 'junit', name: 'junit', version:'4.12'
     testCompile group: 'org.mockito', name: 'mockito-core', version:'2.9.0'
@@ -27,6 +28,8 @@ dependencies {
 
 checkstyle {
     toolVersion = "8.13"
+    //TODO : Create a checkstyle for unit test
+    checkstyleTest.enabled = false
 }
 
 checkstyleMain {
