+++ src/main/java/com/rainbowpunch/jetedge/core/exception/ValueGenerationException.java
@@ -1,5 +1,8 @@
 package com.rainbowpunch.jetedge.core.exception;
 
+/**
+ * An exception for when there is an issue generating values from a limiter.
+ */
 public class ValueGenerationException extends RuntimeException {
     public ValueGenerationException() {
         super();
@@ -17,7 +20,8 @@ public class ValueGenerationException extends RuntimeException {
         super(cause);
     }
 
-    protected ValueGenerationException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+    protected ValueGenerationException(String message, Throwable cause, boolean enableSuppression,
+                                       boolean writableStackTrace) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
 }
