+++ src/main/java/com/rainbowpunch/jetedge/core/limiters/PojoGeneratorLimiter.java
@@ -5,6 +5,10 @@ import com.rainbowpunch.jetedge.spi.pojo.PojoGenerator;
 import java.util.Random;
 import java.util.function.Supplier;
 
+/**
+ * This limiter is a wrapper for existing PojoGenerators so that they can be used as part of a large Limiter.
+ * @param <T>
+ */
 public class PojoGeneratorLimiter<T> extends SimpleAbstractLimiter<T> {
 
     private PojoGenerator<T> generator;
