+++ src/main/java/com/rainbowpunch/jetedge/core/reflection/ConstructorParameter.java
@@ -1,7 +1,13 @@
 package com.rainbowpunch.jetedge.core.reflection;
 
+import lombok.Getter;
+
 import java.util.function.Supplier;
 
+/**
+ * A class for containing the constructor properties to be used in generating end classes.
+ */
+@Getter
 public class ConstructorParameter {
 
     private Supplier<?> objectSupplier;
@@ -10,13 +16,5 @@ public class ConstructorParameter {
     public ConstructorParameter(Object object) {
         objectSupplier = () -> object;
         objectType = object.getClass();
-    }
-
-    public Supplier<?> getObjectSupplier() {
-        return objectSupplier;
-    }
-
-    public Class<?> getObjectType() {
-        return objectType;
     }
 }
